improve organization of react module
