+++ packages/react/src/context.ts
@@ -4,8 +4,7 @@ import { createContext, useContext } from 'react';
 export const LookupContext = createContext(new Context());
 export const useLookup = () => useContext(LookupContext);
 
-export const Pending = new WeakMap<{}, ((context: Context) => void)[]>();
-
+const Pending = new WeakMap<{}, ((context: Context) => void)[]>();
 const Applied = new WeakMap<Model, boolean>();
 
 export function usePeerContext(instance: Model){
@@ -25,4 +24,26 @@ export function usePeerContext(instance: Model){
       Applied.set(instance, true);
     }
   }
+}
+
+export function getPeerContext<T extends Model>(
+  type: Model.Type<T>,
+  relativeTo: Model,
+  callback: (got?: T) => void
+){
+  let pending = Pending.get(relativeTo);
+    
+  if(!pending)
+    Pending.set(relativeTo, pending = []);
+
+  pending.push(context => {
+    callback(context.get(type));
+  })
+}
+
+export function setPeerContext(model: Model, context: Context){
+  const pending = Pending.get(model);
+
+  if(pending)
+    pending.forEach(cb => cb(context));
 }
